rename classes and packages
