+++ pom.xml
@@ -8,11 +8,14 @@
 		<version>2.1.7.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
+
+	<groupId>cn.darkmatter-stream</groupId>
 	<artifactId>darkmatter-stream</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>demo</name>
-	<description>Demo project for Spring Boot</description>
+	<name>Darkmatter Stream</name>
+	<description>Darkmatter Stream</description>
+
+	<packaging>war</packaging>
 
 	<properties>
 		<java.version>1.8</java.version>
@@ -23,13 +26,17 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-	</dependencies>
+  </dependencies>
 
 	<build>
 		<plugins>
