+++ src/main/java/cn/darkmatter/stream/StreamApplication.java
@@ -1,13 +1,13 @@
-package com.example.darkmatterstream;
+package cn.darkmatter.stream;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class StreamApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(StreamApplication.class, args);
 	}
 
 }
