+++ src/main/java/cn/darkmatter/stream/controller/PingController.java
@@ -1,0 +1,15 @@
+package cn.darkmatter.stream.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/ping")
+public class PingController {
+
+    @RequestMapping("")
+    public String ping() {
+        return "good";
+    }
+
+}
