+++ src/main/webapp/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Darkmatter Stream</title>
+</head>
+<body>
+
+<p>Site under construction...</p>
+
+</body>
+</html>
