+++ src/test/java/cn/darkmatter/stream/StreamApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.darkmatterstream;
+package cn.darkmatter.stream;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class DemoApplicationTests {
+public class StreamApplicationTests {
 
 	@Test
 	public void contextLoads() {
